feat(interfaces): create IDataObject.php
The IDataObject interface defines a set of methods that can be implemented by a class to interact with a database using the PHP Data Objects (PDO) extension. The methods defined in this interface cover a range of common database operations such as executing SQL queries, managing transactions, and obtaining information about the database connection.